# scheduler: performance improvement on PodAffinity