@bsalamat 2019-04-09
Performance improvement is over 2X for preferred affinity. This is worth noting in the release notes.