@Huang-Wei 2019-04-09
@bsalamat Will update the release note.

One thing to note is that the 2X improvement is not for soft pod affinity, it's for **hard** pod affinity.

BTW: performance for soft(preferred) pod (anti-) affinity hasn't been measured due to lack of benchmark tests. But from the code's perspective, it's expected to have a performance improvement as well.

Why the priority changes can impact hard pod affinity is because of `hardPodAffinitySymmetricWeight`:

https://github.com/kubernetes/kubernetes/blob/552d1ebd7086dfd4ba5fdc73277866eb656eea0f/pkg/scheduler/algorithm/priorities/interpod_affinity.go#L164-L177