@bsalamat 2019-04-09
Thanks, @Huang-Wei for clarifying. You are right. This improvement impacts both soft and hard affinity.