@Huang-Wei 2019-04-09
Yeap. And hard pod anti-affinity isn't involved in Priorities, so this PR doesn't help for that case.

BTW: I'm also trying to improve hard pod anti-affinity, such as:

1. change podsWithAffinity from slice to set 

https://github.com/kubernetes/kubernetes/blob/c4f97fa2a3a6f952cd2e0d93e237482990b13fe7/pkg/scheduler/nodeinfo/node_info.go#L52

2. split podsWithAffinity into podsWithAffinity and podsWithHardAntiAffinity, so as to short circuit:

https://github.com/kubernetes/kubernetes/blob/c4f97fa2a3a6f952cd2e0d93e237482990b13fe7/pkg/scheduler/algorithm/predicates/metadata.go#L402

But I don't see either of them gains performance improvement yet. Will continue digging.