@bsalamat 2019-04-10
One thing that can improve performance, is to remove `podsWithAffinity []*v1.Pod` from the `NodeInfo` and make it independent (or maybe a part of metadata). Today, we go over the whole "NodeInfo" array and check if any of them has a non-zero length `podWithAffinity`. In clusters with only a few affinity pods,  `podWithAffinity` is empty for most of the "NodeInfo" entries. If `podWithAffinity` was outside of NodeInfo, we could populate the entries only for those nodes which had a pod with affinity.

something like:
```go
NodesWithAffinityPods map[string]podsWithAffinity     // map from node name to affinity pods
```