@Huang-Wei 2019-04-11
FYI: after merge of this PR, the perf improvement can be observed in perf.k8s.io:

![image](https://user-images.githubusercontent.com/1425903/56002391-13872b80-5c77-11e9-9c23-a6d728ae3517.png)