@bsalamat 2019-04-12
@Huang-Wei Scheduling benchmarks show faster execution, but our Kubemark and real cluster benchmarks show a regression:
* Kubemark: http://perf-dash.k8s.io/#/?jobname=kubemark-5000Nodes&metriccategoryname=Scheduler&metricname=SchedulingLatency&Operation=priority_evaluation
* Real cluster benchmark: http://perf-dash.k8s.io/#/?jobname=gce-5000Nodes&metriccategoryname=Scheduler&metricname=SchedulingThroughput